Session-1/MNIST/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.1.1
aom=3.6.0=h313beb8_0
blas=1.0=openblas
bottleneck=1.4.2=py39hbda83bc_0
brotli-python=1.0.9=py39h313beb8_9
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.7.14=hbd8a1cb_0
cairo=1.16.0=hd886ae0_6
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=py39h7f933ea_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
contourpy=1.2.1=py39h48ca7d4_1
cpython=3.9.23=py39hd8ed1ab_0
cycler=0.11.0=pyhd3eb1b0_0
dav1d=1.2.1=h80987f9_0
expat=2.7.1=h313beb8_0
filelock=3.18.0=pyhd8ed1ab_0
fontconfig=2.14.1=h6402c1e_3
fonttools=4.55.3=py39h80987f9_0
freetype=2.13.3=h47d26ad_0
fribidi=1.0.10=h1a28f6b_0
fsspec=2025.7.0=pyhd8ed1ab_0
future=1.0.0=pyhd8ed1ab_2
gettext=0.21.0=hbdbcc25_2
gmp=6.3.0=h7bae524_2
gmpy2=2.2.1=py39h478d0be_0
graphite2=1.3.14=hc377ac9_1
h2=4.2.0=pyhd8ed1ab_0
harfbuzz=10.2.0=he637ebf_1
hpack=4.1.0=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=73.1=h313beb8_0
idna=3.10=pyhd8ed1ab_1
importlib_resources=6.4.0=py39hca03da5_0
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.4.2=py39hca03da5_0
jpeg=9e=h80987f9_3
kiwisolver=1.4.4=py39h313beb8_0
lcms2=2.16=he26ebf3_1
lerc=4.0.0=h313beb8_0
libabseil=20240116.2=cxx17_h00cdb27_1
libavif=1.1.1=h80987f9_0
libblas=3.9.0=1_h2ec9a88_netlib
libcblas=3.9.0=5_h880f123_netlib
libcxx=17.0.6=he5c5206_4
libdeflate=1.22=h80987f9_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.84.2=hdc2269c_0
libiconv=1.16=h80987f9_3
liblapack=3.9.0=5_h880f123_netlib
libopenblas=0.3.29=hea593b9_0
libpng=1.6.39=h80987f9_0
libprotobuf=4.25.3=h514c7bf_0
libtiff=4.7.0=h91aec0a_0
libtorch=2.5.1=gpu_mps_h82d5d13_202
libuv=1.51.0=h5505292_0
libwebp-base=1.3.2=h80987f9_1
libxml2=2.13.8=h0b34f26_0
llvm-openmp=17.0.6=h2f17746_0
lz4-c=1.9.4=h313beb8_1
markupsafe=3.0.2=py39hefdd603_1
matplotlib=3.9.2=py39hca03da5_1
matplotlib-base=3.9.2=py39h7ef442a_1
mpc=1.3.1=h8f1351a_1
mpfr=4.2.1=hb693164_3
mpmath=1.3.0=pyhd8ed1ab_1
ncurses=6.4=h313beb8_0
networkx=3.2.1=pyhd8ed1ab_0
ninja=1.12.1=h420ef59_0
nomkl=1.0=h5ca1d4c_0
numexpr=2.10.1=py39h5d9532f_0
numpy=1.26.4=py39h3b2db8e_0
numpy-base=1.26.4=py39ha9811e2_0
openjpeg=2.5.2=hba36e21_1
openssl=3.5.1=h81ee809_0
packaging=24.2=py39hca03da5_0
pandas=2.3.1=py39h43d6402_0
pcre2=10.42=hb066dcc_1
pillow=11.3.0=py39h6dd570d_0
pip=25.1=pyhc872135_2
pixman=0.40.0=h1a28f6b_0
pybind11-abi=4=hd3eb1b0_1
pycparser=2.22=pyh29332c3_1
pyparsing=3.2.0=py39hca03da5_0
pysocks=1.7.1=pyha55dd90_7
python=3.9.7=hc0da0df_3_cpython
python-dateutil=2.9.0post0=py39hca03da5_2
python-tzdata=2025.2=pyhd3eb1b0_0
python_abi=3.9=8_cp39
pytorch=1.12.1=cpu_py39h6ba7f14_0
pytz=2025.2=py39hca03da5_0
pyyaml=6.0.2=py39hefdd603_2
readline=8.2=h1a28f6b_0
requests=2.32.4=pyhd8ed1ab_0
scikit-learn=1.6.1=py39h313beb8_0
scipy=1.13.1=py39hd336fd7_1
seaborn=0.13.2=py39hca03da5_3
setuptools=78.1.1=py39hca03da5_0
six=1.17.0=py39hca03da5_0
sleef=3.7=h7783ee8_0
sqlite=3.50.2=h79febb2_1
sympy=1.14.0=pyh2585a3b_105
threadpoolctl=3.5.0=py39h33ce5c2_0
tk=8.6.14=h6ba3021_1
torchvision=0.13.0=cpu_py39h80956b8_0
tornado=6.5.1=py39h80987f9_0
tqdm=4.67.1=py39h33ce5c2_0
typing-extensions=4.14.1=h4440ef1_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
unicodedata2=15.1.0=py39h80987f9_1
urllib3=2.5.0=pyhd8ed1ab_0
wheel=0.45.1=py39hca03da5_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h3422bc3_2
zipp=3.21.0=py39hca03da5_0
zlib=1.2.13=h18a0788_1
zstandard=0.23.0=py39hf3bc14e_2
zstd=1.5.6=hfb09047_0
